Фикс мелких багов

+++ src/PHP/DelAdmin.php
@@ -4,8 +4,10 @@ $mysql = new mysqli('localhost', 'root', 'root', 'regist');
 $id = $_GET['id'];
 $category = $_GET['cat'];
 $picture = $_GET['picture'];
-$link = "../../assets/images/$category/$picture";
-unlink($link);
+if($picture != NULL) {
+    $link = "../../assets/images/$category/$picture";
+    unlink($link);
+}
 $mysql->query("DELETE FROM `$category` WHERE `$category`.`id` = '$id'");
 header('Location: ../../pages/admin.php');
 ?>
